+++ CsBot.Tests/WavingHands/AmnesiaTest.cs
@@ -1,7 +1,7 @@
 using CsBot.Games.WavingHands;
-using CsBot.Games.WavingHands.Enchantments;
 using CsBot.Games.WavingHands.Living;
 using CsBot.Games.WavingHands.Spells;
+using CsBot.Games.WavingHands.Spells.Enchantments;
 using NUnit.Framework;
 
 namespace CsBot.Tests.WavingHands
