+++ CsBot.Tests/WavingHands/CureLightWoundsTest.cs
@@ -1,5 +1,6 @@
 using CsBot.Games.WavingHands;
 using CsBot.Games.WavingHands.Spells;
+using CsBot.Games.WavingHands.Spells.Protections;
 using NUnit.Framework;
 
 namespace CsBot.Tests.WavingHands
