--- src/CsBot/Games/WavingHands/Enchantments/Enchantment.cs
@@ -1,8 +1,0 @@
-using CsBot.Games.WavingHands.Spells;
-
-namespace CsBot.Games.WavingHands.Enchantments
-{
-	public abstract class Enchantment : Spell
-	{
-	}
-}
