+++ src/CsBot/Games/WavingHands/Spells/Enchantments/Enchantment.cs
@@ -1,0 +1,6 @@
+namespace CsBot.Games.WavingHands.Spells.Enchantments
+{
+	public abstract class Enchantment : Spell
+	{
+	}
+}
